refactor: update all imports to FSD structure
Eliminate old `@/` imports and update to `~/` aliases.

#VERCEL_SKIP
