+++ src/shared/components/skill-visualization.tsx
