+++ widgets/hero-section.tsx
@@ -1,12 +1,16 @@
-import { Button } from '@/components/ui/button'
-import { ArrowDown, Download, Github, Linkedin, Mail, Sparkles } from 'lucide-react'
-import Link from 'next/link'
-import { siteConfig } from '@/lib/config'
+import { Button } from "@/components/ui/button"
+import { ArrowDown, Download, Github, Linkedin, Mail, Sparkles } from "lucide-react"
+import Link from "next/link"
+import { siteConfig } from "~/shared/config/site"
 
 export function HeroSection() {
   // Get initials from name
   const getInitials = (name: string) => {
-    return name.split(' ').map(n => n[0]).join('').toUpperCase()
+    return name
+      .split(" ")
+      .map((n) => n[0])
+      .join("")
+      .toUpperCase()
   }
 
   return (
@@ -17,7 +21,7 @@ export function HeroSection() {
         <div className="absolute bottom-20 right-20 w-96 h-96 gradient-mesh-alt rounded-full opacity-20 blur-3xl floating-delayed"></div>
         <div className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-[800px] h-[800px] gradient-mesh rounded-full opacity-10 blur-3xl"></div>
       </div>
-      
+
       <div className="container px-4 md:px-6 relative z-10">
         <div className="flex flex-col items-center space-y-8 text-center">
           {/* Profile image with glass effect */}
@@ -37,27 +41,33 @@ export function HeroSection() {
               <span className="w-2 h-2 bg-green-400 rounded-full mr-2 animate-pulse"></span>
               {siteConfig.tagline}
             </div>
-            
+
             <h1 className="text-4xl font-bold tracking-tighter sm:text-5xl md:text-6xl lg:text-7xl">
-              Hi, I'm{' '}
-              <span className="text-gradient block mt-2">
-                {siteConfig.name}
-              </span>
+              Hi, I'm <span className="text-gradient block mt-2">{siteConfig.name}</span>
             </h1>
-            
+
             <p className="mx-auto max-w-[700px] text-muted-foreground text-lg md:text-xl leading-relaxed">
               {siteConfig.description}
             </p>
           </div>
-          
+
           <div className="flex flex-col sm:flex-row gap-4">
-            <Button size="lg" className="glow-hover bg-gradient-to-r from-purple-600 to-pink-600 hover:from-purple-700 hover:to-pink-700 border-0" asChild>
+            <Button
+              size="lg"
+              className="glow-hover bg-gradient-to-r from-purple-600 to-pink-600 hover:from-purple-700 hover:to-pink-700 border-0"
+              asChild
+            >
               <Link href="#contact">
                 <Mail className="mr-2 h-4 w-4" />
                 Get In Touch
               </Link>
             </Button>
-            <Button size="lg" variant="outline" className="morphism-card border-white/20 hover:bg-white/10" asChild>
+            <Button
+              size="lg"
+              variant="outline"
+              className="morphism-card border-white/20 hover:bg-white/10 bg-transparent"
+              asChild
+            >
               <Link href={siteConfig.resumeUrl} target="_blank">
                 <Download className="mr-2 h-4 w-4" />
                 Download CV
